split up elements
